internal: add integrated completion benchmark
